% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_functions.R
\name{check_system}
\alias{check_system}
\title{Check System Requirements}
\usage{
check_system()
}
\value{
Logical. Returns \code{TRUE} if all requirements are met.
}
\description{
Checks if Node.js, npm, and Staticrypt are available on your system.
}
\details{
This function verifies that Node.js, npm, and Staticrypt are installed
and available in the system PATH. It will attempt to install Staticrypt if it is not found.
}
\examples{
try(check_system())
}
